refactor: move common handler code to a base module
